powerpc: kexec fix for powerpc64
slbia instruction invalidates the SLB, but there was a hang on the first
instruction that caused an SLB miss exception. Doing a full sync after
the slbia causes the SLB to be in a consistent state for the handling of
the SLB exception.

Signed-off by: Benjamin Walsh <benjamin.walsh@windriver.com>