gpio-poweroff: Allow it to work on Raspberry Pi
commit  1c7c11181610a6a2a042df31360364a9a5c967af from
https://github.com/raspberrypi/linux.git rpi-5.17.y

The Raspberry Pi firmware manages the power-down and reboot
process. To do this it installs a pm_power_off handler, causing
the gpio-poweroff module to abort the probe function.

This patch introduces a "force" DT property that overrides that
behaviour, and also adds a DT overlay to enable and control it.

Note that running in an active-low configuration (DT parameter
"active_low") requires a custom dt-blob.bin and probably won't
allow a reboot without switching off, so an external inversion
of the trigger signal may be preferable.
Signed-off-by: Meng Li <Meng.Li@windriver.com>