amba_pl011: Add cts-event-workaround DT property
commit  abe767973bba583d55f840fbab71c4e80b833041 from
https://github.com/raspberrypi/linux.git rpi-5.15.y

The BCM2835 PL011 implementation seems to have a bug that can lead to a
transmission lockup if CTS changes frequently. A workaround was added to
the driver with a vendor-specific flag to enable it, but this flag is
currently not set for ARM implementations.

Add a "cts-event-workaround" property to Pi DTBs and use the presence
of that property to force the flag to be enabled in the driver.

See: https://github.com/raspberrypi/linux/issues/1280

Signed-off-by: Phil Elwell <phil@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>