ARM: Activate FIQs to avoid __irq_startup warnings
commit  937ddc7538062ce5aeb2f0065bcf6c065842b986 from
https://github.com/raspberrypi/linux.git rpi-5.16.y

There is a new test in __irq_startup that the IRQ is activated, which
hasn't been the case for FIQs since they bypass some of the usual setup.

Augment enable_fiq to include a call to irq_activate to avoid the
warning.

Signed-off-by: Phil Elwell <phil@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>