ARM: Activate FIQs to avoid __irq_startup warnings
commit  605bb54a8397aaf31023e4fdd30f9b559534b9fb from
https://github.com/raspberrypi/linux.git rpi-6.2.y

There is a new test in __irq_startup that the IRQ is activated, which
hasn't been the case for FIQs since they bypass some of the usual setup.

Augment enable_fiq to include a call to irq_activate to avoid the
warning.

Signed-off-by: Phil Elwell <phil@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>