net: lan78xx: Disable TCP Segmentation Offload (TSO)
commit  1cdd1094cdb53c06cc6b8b5861ad04c064471e27 from
https://github.com/raspberrypi/linux.git rpi-5.17.y

TSO seems to be having issues when packets are dropped and the
remote end uses Selective Acknowledge (SACK) to denote that
data is missing. The missing data is never resent, so the
connection eventually stalls.

There is a module parameter of enable_tso added to allow
further debugging without forcing a rebuild of the kernel.

https://github.com/raspberrypi/linux/issues/2449
https://github.com/raspberrypi/linux/issues/2482

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>