bcm2835-dma: Add support for per-channel flags
commit  d54375446c4df4f72ceb5f34739bd8d4b40b88c3 from
https://github.com/raspberrypi/linux.git rpi-6.1.y

Add the ability to interpret the high bits of the dreq specifier as
flags to be included in the DMA_CS register. The motivation for this
change is the ability to set the DISDEBUG flag for SD card transfers
to avoid corruption when using the VPU debugger.

Signed-off-by: Phil Elwell <phil@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>