bcm2835-dma: Add support for per-channel flags
commit  5d6d268bad7b1016d55b5e95de086812316245e1 from
https://github.com/raspberrypi/linux.git rpi-6.0.y

Add the ability to interpret the high bits of the dreq specifier as
flags to be included in the DMA_CS register. The motivation for this
change is the ability to set the DISDEBUG flag for SD card transfers
to avoid corruption when using the VPU debugger.

Signed-off-by: Phil Elwell <phil@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>