bcm2835-dma: Add support for per-channel flags
commit  2c63fab1e226c724e5be8d20cfc61167619f6e72 from
https://github.com/raspberrypi/linux.git rpi-5.15.y

Add the ability to interpret the high bits of the dreq specifier as
flags to be included in the DMA_CS register. The motivation for this
change is the ability to set the DISDEBUG flag for SD card transfers
to avoid corruption when using the VPU debugger.

Signed-off-by: Phil Elwell <phil@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>