media: tc358743: Check I2C succeeded during probe.
commit  859b13f9446f0c8ffcb50a5dc526243f80bea317 from
https://github.com/raspberrypi/linux.git rpi-6.0.y

The probe for the TC358743 reads the CHIPID register from
the device and compares it to the expected value of 0.
If the I2C request fails then that also returns 0, so
the driver loads thinking that the device is there.

Generally I2C communications are reliable so there is
limited need to check the return value on every transfer,
therefore only amend the one read during probe to check
for I2C errors.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>