soc: bcm: bcm2835-pm: Add support for 2711.
commit  2a203fe86eb35ec84a52c8f1eef15bff7c497b21 from
https://github.com/raspberrypi/linux.git rpi-5.19.y

Without the actual power management part any more, there's a lot less
to set up for V3D.  We just need to clear the RSTN field for the power
domain, and expose the reset controller for toggling it again.

This is definitely incomplete -- the old ISP and H264 is in the old
bridge, but since we have no consumers of it I've just done the
minimum to get V3D working.

Signed-off-by: Eric Anholt <eric@anholt.net>
Signed-off-by: Meng Li <Meng.Li@windriver.com>