staging: vchiq_arm: Give vchiq children DT nodes
commit  4deb652f208e8329429bf0191306b085d4e34e70 from
https://github.com/raspberrypi/linux.git rpi-5.15.y

vchiq kernel clients are now instantiated as platform drivers rather
than using DT, but the children of the vchiq interface may still
benefit from access to DT properties. Give them the option of a
a sub-node of the vchiq parent for configuration and to allow
them to be disabled.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>