drm/panel/raspberrypi-touchscreen: Use independent I2C actions with delay.
commit  c452a82aca02afe7e443506857b6407444365d9f from
https://github.com/raspberrypi/linux.git rpi-5.16.y

We now have the hardware I2C controller pinmuxed to the drive the
display I2C, but this controller does not support clock stretching.
The Atmel micro-controller in the panel requires clock stretching
to allow it to prepare any data to be read.

Split the rpi_touchscreen_i2c_read into two independent transactions with
a delay between them for the Atmel to prepare the data.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>