media: bcm2835-unicam: Do not stop streaming in unicam_release
commit  10d30e4266253e115caf1b0ebfb2485832fa3d65 from
https://github.com/raspberrypi/linux.git rpi-5.17.y

unicam_release calls _vb2_fop_release, which will call stop_streaming
if that particular node was streaming. Calling it unconditionally (as
the code was) means that if a second handle was opened eg to alter
a setting, on closing that connection it also stopped Unicam.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>