media: bcm2835-unicam: Always service interrupts
commit  5388dcb92643ed95a62814b3293cbe9373a31a71 from
https://github.com/raspberrypi/linux.git rpi-5.19.y

From when bringing up the driver, there was a check in the isr
to ignore interrupts (claiming them handled) should the driver
not be streaming.

The VPU now will not register a camera driver if it finds a
CSI2 node enabled in device tree, therefore this flawed check is
redundant.

https://github.com/raspberrypi/linux/issues/3602

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>