staging:vc04_services: bcm2835-camera: Request headers with I-frame
commit  f523a491ce2f335b131e49971f8ca16a904835f5 from
https://github.com/raspberrypi/linux.git rpi-6.0.y

V4L2 wishes to have the codec header bytes in the same buffer as the
first encoded frame, so it does become 1-in 1-out for encoding.
The firmware now has an option to do this, so enable it.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>