staging:vc04_services: bcm2835-camera: Request headers with I-frame
commit  ef79c89af674f28f4a327a84cce683e8e94a584a from
https://github.com/raspberrypi/linux.git rpi-5.15.y

V4L2 wishes to have the codec header bytes in the same buffer as the
first encoded frame, so it does become 1-in 1-out for encoding.
The firmware now has an option to do this, so enable it.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>