staging: bcm2835-codec: Allow decode res changed before STREAMON(CAPTURE)
commit  996a943e36a45a1cc9a2cdc062435f1317a9609f from
https://github.com/raspberrypi/linux.git rpi-5.17.y

The V4L2 stateful video decoder API requires that you can STREAMON
on only the OUTPUT queue, feed in buffers, and wait for the
SOURCE_CHANGE event.
This requires that we enable the MMAL output port at the same time
as the input port, because the output port is the one that creates
the SOURCE_CHANGED event.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>