staging: bcm2835-codec: Allow decode res changed before STREAMON(CAPTURE)
commit  ec7812707197cc7c58120b525525fc3e8f5c21f2 from
https://github.com/raspberrypi/linux.git rpi-5.18.y

The V4L2 stateful video decoder API requires that you can STREAMON
on only the OUTPUT queue, feed in buffers, and wait for the
SOURCE_CHANGE event.
This requires that we enable the MMAL output port at the same time
as the input port, because the output port is the one that creates
the SOURCE_CHANGED event.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>