firmware: raspberrypi: Add support for tryonce reboot flag
commit  4f3c2c91c7c739efb844bb1601c7c21039fbea41 from
https://github.com/raspberrypi/linux.git rpi-5.17.y

Define a new mailbox (SET_REBOOT_FLAGS) which may be used to
pass optional flags to the Raspberry Pi firmware that changes
the behaviour of the bootloader and firmware during a reboot.

Currently this just defines the 'tryboot' flag which causes
the firmware to load tryboot.txt instead config.txt. This
alternate configuration file can be used to specify the
path of an alternate firmware and kernels allowing a fallback
mechanism to be implemented for OS upgrades.
Signed-off-by: Meng Li <Meng.Li@windriver.com>