bcm2835-isp: Allow formats with different colour spaces.
commit  43dfe4358cf7d8aa77cf67ab7d95291c4ef31450 from
https://github.com/raspberrypi/linux.git rpi-5.18.y

Each supported format now includes a mask showing the allowed colour
spaces, as well as a default colour space for when one was not
specified.

Additionally we translate the colour space to mmal format and pass it
over to the VideoCore.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>