bcm2835-isp: Allow formats with different colour spaces.
commit  02e954291528595eebf1cab58f4f5e94e600e2ca from
https://github.com/raspberrypi/linux.git rpi-6.1.y

Each supported format now includes a mask showing the allowed colour
spaces, as well as a default colour space for when one was not
specified.

Additionally we translate the colour space to mmal format and pass it
over to the VideoCore.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>