drm/vc4: hdmi: Change CSC callback prototype
commit  896b4774d6a4e2fc370624cddd995d046f78139c from
https://github.com/raspberrypi/linux.git rpi-5.17.y

In order to support the YUV output, we'll need the atomic state to know
what is the state of the associated property in the CSC setup callback.

Let's change the prototype of that callback to allow us to access it.

Acked-by: Thomas Zimmermann <tzimmermann@suse.de>
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>