drm/vc4: Change the default DPI format to being 18bpp, not 24.
commit  a515a19159d71f0a5d0671ed80824a2c7c7acb5e from
https://github.com/raspberrypi/linux.git rpi-6.1.y

DPI hasn't really been used up until now, so the default has
been meaningless.
In theory we should be able to pass the desired format for the
adjacent bridge chip through, but framework seems to be missing
for that.

As the main device to use DPI is the VGA666 or Adafruit Kippah,
both of which use RGB666, change the default to being RGB666 instead
of RGB888.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>