media: i2c: imx477: Remove internal v4l2_mbus_framefmt from the state
commit  091dfe802ef937553ed8fd4d605a94b1e0894647 from
https://github.com/raspberrypi/linux.git rpi-5.19.y

The only field in this struct that is used is the format code, so
replace the struct with this single field.

Save the format code in imx477_set_pad_format() when setting up a new
mode so that imx477_get_pad_format() performs the right lookup.
Otherwise, this caused a bug where the mode lookup occurred on the
12-bit table rather than the 10-bit table.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>