staging/bcm2835-camera: Add support for DMABUFs
commit  b7b1ff5d5513118e0dc526ee739a015799c76c05 from
https://github.com/raspberrypi/linux.git rpi-6.1.y

DMABUFs are all handled by videobuf2, so there is no reason not
to enable support for them.

Note that this driver is still using the vmalloc allocator, so
the buffers it allocates will not be compatible with the codec
or ISP driver that require contiguous buffers. However this
driver should be able to import the buffers allocated by them.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>