clk: Always clamp the rounded rate
commit  282be75c616dfe0b6d74936306697a7849845f91 from
https://github.com/raspberrypi/linux.git rpi-5.17.y

The current core while setting the min and max rate properly in the
clk_request structure will not make sure that the requested rate is
within these boundaries, leaving it to each and every driver to make
sure it is.

Add a clamp call to make sure it's always done.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>