media: i2c: imx258: Make V4L2_CID_VBLANK configurable.
commit  f78df58188cc4c2766cdea0984ac5aba5dc6ecbc from
https://github.com/raspberrypi/linux.git rpi-5.18.y

The values and ranges of V4L2_CID_VBLANK are all computed,
so there is no reason for it to be a read only control.
Remove the register values from the mode lists, add the
handler, and remove the read only flag.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>