media: i2c: imx258: Make V4L2_CID_VBLANK configurable.
commit  86331bbcc0d9a50c0460db7a4db982950be311e0 from
https://github.com/raspberrypi/linux.git rpi-5.19.y

The values and ranges of V4L2_CID_VBLANK are all computed,
so there is no reason for it to be a read only control.
Remove the register values from the mode lists, add the
handler, and remove the read only flag.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>