media: i2c: imx258: Make V4L2_CID_VBLANK configurable.
commit  0c34c3a1626448babc70ee5d610699fad8a6545a from
https://github.com/raspberrypi/linux.git rpi-6.0.y

The values and ranges of V4L2_CID_VBLANK are all computed,
so there is no reason for it to be a read only control.
Remove the register values from the mode lists, add the
handler, and remove the read only flag.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>